+++ flask/app/core/secret_santa.py
@@ -2,6 +2,7 @@
 # -*- coding: utf-8 -*-
 import random
 from copy import copy
+from uuid import uuid4
 
 
 def random_number(size):
@@ -10,6 +11,9 @@ def random_number(size):
     return random.randint(0, size-1)
 
 def create_associations(guests:list) -> list:
+    for guest in guests:
+        guest["link"] = str(uuid4())
+
     ended = False
     while not ended:
         guest_done = []
@@ -20,8 +24,7 @@ def create_associations(guests:list) -> list:
             if target in guest_done:
                 break
             if target != guest:
-                guest["target"] = target["name"]
-                guest["target_email"] = target["email"]
+                guest["target_link"] = target["link"]
                 guest_done.append(target)
                 target_tmp.remove(target)
         ended = len(target_tmp) == 0